requirements.txt
# Document Processing
PyMuPDF==1.23.8
python-docx==1.0.1
openpyxl==3.1.2
pdf2image==1.16.3
pytesseract==0.3.10

# Image Processing
torch==2.1.0
transformers
salesforce-lavis
clip==0.2.0

# Table Processing
pandas==2.1.0
# tapas-pytorch==0.0.1

# Vector Store and Retrieval
faiss-cpu==1.7.4
sentence-transformers==2.2.2
langchain==0.0.350
llama-index==0.9.8

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# LLM
zhipuai==1.0.7
dashscope==1.13.6
openai==1.3.0  